CHANGE: TransformerMixin classes organisation
